Adds password reset link to log-in form
