+++ app/brochure/routes/log-in/checkEmail.js
@@ -5,6 +5,8 @@ module.exports = function checkEmail(req, res, next) {
   var email = req.body && req.body.email;
 
   if (!email) return next(new LogInError("NOEMAIL"));
+  
+  res.locals.email = email;
 
   User.getByEmail(email, function(err, user) {
     if (err) return next(err);
@@ -19,8 +21,9 @@ module.exports = function checkEmail(req, res, next) {
     // can delete this check safely.
     if (user.isDisabled) return res.redirect('/account/disabled')
 
-    req.user = user;
     res.locals.email = user.email;
+    res.locals.validemail = true;
+    req.user = user;
     res.locals.then = req.query.then;
 
     next();
