--- app/brochure/views/log-in/password.html
@@ -1,50 +1,0 @@
-<style type="text/css">
-input[type="submit"].link {    font: inherit;
-    color: inherit;
-    background: none;
-    border: none;
-    padding: 0;
-    display: inline;
-    margin: 0;
-    color: #005eb8;
-    text-decoration: underline;
-    transition-property: color, border-color;
-    transition-duration: .1s;
-    transition-timing-function: ease-out;}
-    input[type="submit"].link:hover {border-bottom-color: #111}
-</style>
-
-
-<form method="POST" style="margin:0;padding:0">
-
-  <input type="hidden" name="email" value="{{email}}" />
-
-  {{^error}}
-  <p>Please enter your password below or <input type="submit" class="link" name="reset" value="reset your password"> if you’ve forgotten it.</p>
-  {{/error}}
-
-  {{#NOPASSWORD}}
-  <p class="error">
-    Please enter a password. Would you like to <input type="submit" class="link" name="reset" value="reset your password">?
-  </p>
-  {{/NOPASSWORD}}
-
-  {{#BADPASSWORD}}
-  <p class="error">
-    That was not your password. Would you like to <input type="submit" class="link" name="reset" value="reset your password">?
-  </p>
-  {{/BADPASSWORD}}
-
-</form>
-
-<form method="POST">
-  <p>
-    <label for="email">Your email address</label>
-    <input name="email" type="text" value="{{email}}" />
-  </p>
-    <p>
-    <label for="password">Password</label>
-    <input name="password" type="password" autofocus />
-  </p>
-  <input type="submit" value="Log in" />
-</form>
